python/requirements.txt
# [dev-packages]
pytest
flake8
black

# [packages]
PyYAML
matplotlib
